partial fix(generator): use correct TS array type for method parameters (#313)
* partial fix(generator): use correct TS array type for method parameters

`array` is not a valid type in TS. At the very least, it should be either `Array<any>` or `any[]`, and preferably, it should also include the type information of the array items (e.g. `Array<number>` or `number[]`).

This is a partial fix for #312. It only fixes the generated type of method parameters (generate `any[]` instead of `array`). The type of the return values is still generated wrongly.

Related to #312